+++ package.json
@@ -11,12 +11,11 @@
     "url": "https://github.com/htmlacademy-javascript/563053-kekstagram-28/issues"
   },
   "devDependencies": {
-    "browser-sync": "2.27.11",
+    "browser-sync": "^2.28.3",
     "eslint": "8.33.0",
     "eslint-config-htmlacademy": "9.0.0"
   },
   "license": "MIT",
-  "dependencies": {},
   "scripts": {
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "lint": "eslint js/"
